requirements.txt
actionlib==1.13.2
angles==1.9.13
appdirs==1.4.3
apturl==0.5.2
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
catkin-tools==0.9.2
cbor==1.0.0
ccsm==0.9.14.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.2.0
Click==7.0
cmake==3.27.0
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
constantly==15.1.0
construct==2.8.8
contourpy==1.1.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.3
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
fastimport==0.9.8
filelock==3.0.12
fonttools==4.41.1
future==0.18.2
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
geographiclib==2.0
geopy==2.3.0
gpg===1.13.1-unknown
h5py==3.7.0
hexdump==3.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib-metadata==1.5.0
importlib-resources==6.0.0
incremental==16.10.1
interactive-markers==1.12.0
Jinja2==3.1.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.6
lockfile==0.12.2
logutils==0.3.5
louis==3.12.0
lxml==4.9.1
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.3
matplotlib==3.7.2
message-filters==1.15.14
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
nose==1.3.7
ntplib==0.4.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
opencv-python==4.6.0.66
osef==2.4.0
osrf-pycommon==2.0.2
packaging==21.3
pandas==2.0.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipreqs==0.4.13
pr2-power-board==1.1.10
protobuf==3.6.1
psutil==5.5.1
py==1.11.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
pypylon==2.2.0
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyshark==0.5.3
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2023.3
pyusb==1.2.1
pyxdg==0.26
PyYAML==5.3.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
rainbow-logging-handler==2.2.2
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rosapi==0.11.16
rosbag==1.15.14
rosboost-cfg==1.15.8
rosbridge-library==0.11.16
rosbridge-server==0.11.16
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.22.1
rosdep-modules==0.22.1
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosgraph==1.15.14
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.15.14
roslib==1.15.8
roslibpy==1.3.0
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospkg-modules==1.4.0
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-pr2-dashboard==0.4.0
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.20
scipy==1.10.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
style==1.1.0
sympy==1.12
systemd-python==234
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.6
tk==0.1.0
topic-tools==1.15.14
torch==2.0.1
torchvision==0.15.2
tornado==5.1.1
tqdm==4.65.0
triton==2.0.0
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.7.1
tzdata==2023.3
u-msgpack-python==2.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
ultralytics==8.0.141
unattended-upgrades==0.1
update==0.0.1
urllib3==1.25.8
usb-creator==0.3.7
vcstools==0.1.42
virtualenv==20.0.17
wadllib==1.3.3
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.13
xkit==0.0.0
yarg==0.1.9
zipp==3.16.2
zope.interface==4.7.1
